+++ utils/challenges/Challenge4.ts
@@ -8,6 +8,8 @@ import BotClient from '../../utils/classes/BotClient';
 
 import SolvedCaptchaResult from '../interfaces/SolvedCaptchaResult';
 
+import { captchaModal } from '../../commands/group/general/login';
+
 function getInstruction(challenge: Challenge4) {
     let instructionString = (challenge.data.game_data as any).instruction_string;
     let instruction = challenge.data.string_table[`4.instructions-${instructionString}`];
@@ -17,31 +19,37 @@ function getInstruction(challenge: Challenge4) {
 }
 
 export default async function solveChallenge4(interaction: Discord.CommandInteraction, client: BotClient, challenge: Challenge4): Promise<SolvedCaptchaResult> {
+    const buttonFilter = (buttonInteraction: Discord.Interaction) => buttonInteraction.isButton() && buttonInteraction.user.id === interaction.user.id;
+    const modalFilter = (modalInteraction: Discord.ModalSubmitInteraction) => modalInteraction.isModalSubmit() && modalInteraction.user.id === interaction.user.id;
     try {
         let amountOfWaves = challenge.data.game_data.waves;
         let embed = client.embedMaker({title: "Captcha Required", description: `Logins require a captcha to be completed, please complete the captcha below\n\nObjective: ${getInstruction(challenge)}\n\nGuide: https://i.imgur.com/05OYegq.png\n\nAmount of Waves: ${amountOfWaves}\n\nTo answer, type the # tile that the answer is in`, type: "info", author: interaction.user});
+        let componentData = client.createButtons([
+            {customID: "Answer", label: "Answer", style: Discord.ButtonStyle.Primary}
+        ]);
         await interaction.editReply({embeds: [embed]});
         for(let i = 0; i < amountOfWaves; i++) {
             await fs.promises.writeFile(`${process.cwd()}/Image.gif`, await challenge.getImage());
-            let msg = await (interaction.channel as Discord.TextChannel).send({files: [`${process.cwd()}/Image.gif`]});
-            let collected = await interaction.channel.awaitMessages({
-                filter: (message: Discord.Message) => {
-                    if(message.author.id !== interaction.user.id) return false;
-                    return true;
-                },
-                time: client.config.collectorTime,
-                max: 1
-            });
-            await msg.delete();
-            if(collected.size === 0) {
+            let msg = await (interaction.channel as Discord.TextChannel).send({files: [`${process.cwd()}/Image.gif`], components: componentData.components});
+            let button = (await msg.awaitMessageComponent({filter: buttonFilter, time: client.config.collectorTime}));
+            if(!button) {
+                await msg.delete();
                 let embed = client.embedMaker({title: "Captcha Expired", description: "You didn't answer the captcha in time, please rerun the command", type: "error", author: interaction.user});
                 await interaction.editReply({embeds: [embed]});
                 return {success: false, error: "CE"};
             }
-            try {
-                await collected.at(0).delete();
-            } catch {}
-            let answer = Number(collected.at(0).content);
+            await button.showModal(captchaModal);
+            let modal = await button.awaitModalSubmit({filter: modalFilter, time: client.config.collectorTime});
+            if(!modal) {
+                await msg.delete();
+                let embed = client.embedMaker({title: "Captcha Expired", description: "You didn't answer the captcha in time, please rerun the command", type: "error", author: interaction.user});
+                await interaction.editReply({embeds: [embed]});
+                return {success: false, error: "CE"};
+            }
+            await msg.delete();
+            await modal.reply({content: "ㅤ"});
+            await modal.deleteReply();
+            let answer = Number(modal.fields.getTextInputValue("answer"));
             if(isNaN(answer)) {
                 return {success: false, error: "Invalid answer received"};
             }
